Merged test instances into test steps. Now test steps are the only supported mechanism for representing tests created at runtime.  This is good because it affords us some simplification of the data model.  I originally avoided doing this very thing because I thought I needed to be able to distinguish these cases.  It turns out that it was sufficient to just introduce the concept of a "Primary" test step.
Fixed a problem with the model's ability to express skipped tests and tests that are aborted before they get a chance to begin execution.  Previously until the test instance reached its Initialize phase, there was no context for it.  So any log output would be recorded against the parent test.  That resulted in some rather strange looking output.  Now instead we create a primary test step as soon as possible.  If the test is not data-driven, then we reuse the primary test step for the rest of the run.  However, if it is data-driven then we create derived test steps as children of the primary test step.

One side-effect of these changes is that data-driven tests are now grouped together under their primary test in the reports.  I think most users will appreciate this feature anyways.

Also changed the report XSLT to use solid colour progress bars by default instead of showing proportional results.  This assists with readability in reports with large numbers of tests.

Made a few other minor tweaks to presentation here and there such as in TD.Net message output and in the text reports.  Also updated the NBehave prototype to bring it in sync with the API changes.