Refactored the pattern construction mechanism to use a suite of builders instead of directly modifying pattern test objects.  This is just one step towards the upcoming goal of eliminating the test tree as it currently exists to adopt a more lightweight specification.
Added an [XmlData] attribute and fixed numerous issues in the Xml data set.

Added stubs for the mixin feature, not yet functional.

Refactored and greatly reduced the complexity of the contract verifier feature.  This takes the idea of field-based contract verifiers to their logical extension: they are now constructed entirely dynamically.  The pipeline reuses portions of the hierarchy of Test types also used for test factories.  The extra pattern, builder and settings classes have been eliminated with the remainded consolidated and folded into the contracts which themselves are more reusable than they once were.