Changed the test runner lifecycle to just Explore & Run, where each encapsulates the full process end-to-end.  The benefit here is that the AggregateTestDriver no longer needs to keep all of the partitions active simultaneously.  As a result, overall memory consumption is significantly reduced when running tests from large numbers of test assemblies.
The whole point here was to finally workaround a huge performance problem that recently has been causing OutOfMemoryExceptions on the build server.

Added a debugger API and a debugging flag in HostSetup.  Moved all of the debugging support code from Icarus into Gallio core.

Added /debug flag to Echo and equivalent options to NAnt, MSBuild and PowerShell.

Improved how log messages from the host process are presented in the caller by encoding the message severity as part of the output text.

Split out the Visual Studio interop components into a separately install (but at the moment mandatory) feature.  Changed where the Navigator component gets installed so it uses to the dll in the root Gallio bin folder now.

Adopted a notification pipeline for exporting the test model out of the test driver.  This starts paving the way towards a proper asynchronous messaging infrastructure.

Deleted a ton of unused code that was created for Gallio.Sail.  Parts of it will go back in later as needed, presumably...

Work in progress... Icarus in particular is probably broken in a dozen ways due to the amount of rework required to bring it in line with the new lifecycle.  (Sorry Graham!)