Removed an excess dependency that was causing some tests to fail.
The R# Psi reflection policy returns all references assemblies in the project as assembly references but the compiler is smarter than that and will include only referenced assemblies that are actually used.  So Psi will find System.Data whereas native reflection will not because it is not actually used by MbUnit.TestResources.