Added an ErrorDialog to be shared and modified all of the tools to use it instead of MessageBox.
Fixed a minor issue with MultipleCultureAttribute, RepeatAttribute and ThreadedRepeatAttribute where the nested steps were being created with the wrong code element.  This meant that the user could not navigate from the step back to the test.

Removed the settings page from the installer.  Added a preference pane for the Ambience server instead.

Fixed a bug in the TDNet preference pane where the framework settings were getting mixed up because of a mistaken assumption about the order of the rows.