Added a workaround for the IE Local Machine Zone Lockdown so that links to attachments in local files work.  The workaround detects when the user is viewing a locally stored report in IE and updates attachment links to point to resources provided by the gallio pluggable protocol handler.
Some of the code in the Gallio Navigator project is a bit overcomplicated now because it took me a awhile to figure out why it was not working.  I'm leaving it in there to support future feature expansion if needed.

Added some more general catch blocks in a few places where we were previously catching IOException only because it is not enough.  We can also see a bunch of unrelated exceptions like UnauthorizedAccessException that break recovery logic downstream.

Added support for detecting mime types of files by extension using the Windows registry when otherwise unknown.

Reenabled the maximum image size limit in reports now that attachment links work in IE again.  This way reports don't get too wide but the user can still click on each image to view it in full size.