Rewrote ShellEnvironment to remove the inadvertent dependency that was introduced in the critical load path of the Shell package and add-in.  It is not possible to load Gallio.VisualStudio.Shell.dll until the runtime is initialized.
Added a workaround for an NRE when a R# IModule cannot be found from which to read custom attributes.

Take advantage the new TaskException constructor in R# 5.0 in the TaskExceptionFactory class.  No more ugly reflection hacks needed.

Found a bug in the design of R#'s RemoteTask class implementation of IEquatable<T> that was probably responsible for popping up spurious message boxes about test failures as if they were assembly load errors.  Added some salt to reduce the likelihood of hash collisions in the meantime.